Reorg: Move api into new kubernetes_api crate
kubernetes_api is everything that serialises to/from the Kubernetes
(HTTP/JSON) wire protocol.

Changes beyond naive rename:
- Moved `Metadata`, `List`, `ItemList` to `meta::v1::*`
- Renamed `TypeMetaStruct` -> `TypeMetaImpl`
- Moved `GroupVersion` (and friends) to `meta::GroupVersion`
- Moved `Unstructured` into `unstructured` sub-module
- Add `workspace` declaration to top-level `Cargo.toml`
