--- holding/src/api/meta/mod.rs
@@ -1,1 +1,0 @@
-pub mod v1;
